signed short

+++ blocks/GLWindow.h
@@ -20,7 +20,7 @@ public:
 	void ReSizeGLScene( GLsizei width, GLsizei height );
 	bool CreateGLWindow( LPCSTR title, GLsizei width, GLsizei height, int bits);
 	int DrawGLScene();
-	void GlTile(signed long X, signed long Y, signed long Z, char N);
+	void GlTile(signed short X, signed short Y, signed short Z, char N);
 
 	GLsizei width, height;
 
@@ -32,11 +32,11 @@ public:
 	bool  keys[256];				// ћассив, используемый дл€ операций с клавиатурой
 	bool  active;					// ‘лаг активности окна, установленный в true по умолчанию
 
-	Tile* FindTile(signed long x, signed long y, signed long z);
-	void FindTileN(signed long x, signed long y, signed long z, char N);
-	int AddTile(signed long x, signed long y, signed long z, char mat);
-	int RmTile(signed long x, signed long y, signed long z);
-	signed long Hash(signed long x, signed long y, signed long z);
+	Tile* FindTile(signed short x, signed short y, signed short z);
+	void FindTileN(signed short x, signed short y, signed short z, char N);
+	int AddTile(signed short x, signed short y, signed short z, char mat);
+	int RmTile(signed short x, signed short y, signed short z);
+	unsigned long Hash(signed short x, signed short y, signed short z);
 	Tiles *tTiles;
 
 	std::deque<Tile *> *visible;
